Migrate project to ESM
