+++ package.json
@@ -13,6 +13,7 @@
     "postcss-rtlcss",
     "rtlcss"
   ],
+  "type": "module",
   "main": "index.js",
   "module": "esm/index.js",
   "types": "index.d.ts",
@@ -52,7 +53,7 @@
     "clean": "./scripts/clean.sh",
     "copy": "./scripts/copy.sh",
     "modify-dts": "replace-in-file --configFile=config.replace.json",
-    "build": "pnpm clean && rollup --config rollup.config.js --bundleConfigAsCjs && pnpm copy && pnpm modify-dts",
+    "build": "pnpm clean && rollup --config rollup.config.mjs && pnpm copy && pnpm modify-dts",
     "playground:start": "pnpm --filter postcss-rtlcss-playground start",
     "playground:build": "pnpm --filter postcss-rtlcss-playground build",
     "prepare": "pnpm build",
@@ -85,7 +86,7 @@
     "jest": "^30.4.2",
     "jest-specific-snapshot": "^8.0.0",
     "postcss": "^8.4.38",
-    "postcss-preset-env": "10.6.1",
+    "postcss-preset-env": "11.3.2",
     "replace-in-file": "^8.4.0",
     "rimraf": "^6.1.3",
     "rollup": "^4.62.4",
