+++ playground/aliases.js
@@ -1,10 +1,10 @@
-const path = require('path');
-const tsconfig = require('./tsconfig');
+import path from 'node:path';
+import tsconfig from './tsconfig.json' with { type: "json" };
 
 const { compilerOptions: { paths } } = tsconfig;
 const aliasReg = (str) => str.replace(/^(.*)\/\*$/, '$1');
 
-module.exports = Object.keys(paths).reduce(
-    (obj, a) => (obj[aliasReg(a)] = path.resolve(__dirname, aliasReg(paths[a][0])), obj),
+export default Object.keys(paths).reduce(
+    (obj, a) => (obj[aliasReg(a)] = path.resolve(aliasReg(paths[a][0])), obj),
     {}
 );
