+++ scripts/copy.sh
@@ -3,6 +3,3 @@
 mkdir esm
 
 cp -a dist/. ./
-
-## esm package
-echo '{\n    "type": "module"\n}' > esm/package.json
